feat(prototype): compute shearzone-borehole coords
Compute global coordinates of shearzone-borehole
intersections.
The shearzones are fitted with methods in fit_plane.
Use the unit normal and mean coordinate to compute
"new" intersection with each borehole.
Only boreholes in the shearzone-borehole dataset is
considered.
I.e. intersections that aren't used in the regression
(but may be classified in the borehole-shearzone dataset)
will not be considered.
Firstly: We do not know which specific shear-zone these
intersections belong to (only the set).

fixes subproblem in #5