feat(BREAKING): Make model setup more abstract
By utilizing the fact that methods can be used
as parameters for other functions, we are able
to make the setup of - and running models more
general (and hopefully more flexible).
I.e.: The top-level methods are much shorter
because the specifics (or default values)
are delegated to specialized methods.